USBPhy_STM32: Allow for endpoint max packets sizes greater than 64.
The MPS for USB HS bulk endpoints is 512 bytes.

USBPhy_STM32.cpp makes some attempt to support USB HS but it hasn't been correctly (or fully). The 64 byte MPS is a FS thing.

This doesn't mean USBPhy_STM32.cpp is now fully USB HS capable but it is a step in the right direction.